Add utility to share rust reference across Python objects
This was originally developed for the Mercurial's Rust extension. It allows
us to implement a Python iterator over a Rust iterator _relatively_ safely.

The problem is that a Rust iterator typically has a reference to the
collection behind, and thus cannot be a data member of a Python object.
We get around this problem by casting &'a to &'static and add runtime
validation instead. The basic idea is described in the following blog post.

https://raphaelgomes.dev/blog/articles/2019-07-01-sharing-references-between-python-and-rust.html

In order to make the reference-sharing business safe, we have the following
structs:

 - PySharedRefCell defines a storage holding a value to be shared
 - PySharedRef is a lifetime-bound reference to the value above, which provides
   a safe interface on top of the PySharedRefCell and its owner PyObject
 - UnsafePyLeaked is the reference not bound to the real lifetime, and its
   validity is checked at runtime
 - PyLeakedRef/PyLeakedRefMut are borrowed references from UnsafePyLeaked

In order to create PySharedRef in safe manner, py_class!() macro extension
will be added by later patches:

    py_class!(class List |py|
        @shared data vec: Vec<i32>;
        // the storage type is PySharedRefCell<Vec<i32>>, but its accessor
        // returns PySharedRef<'a, Vec<i32>>. It makes sure that the right
        // owner PyObject and its data member are paired.
    });

There are a couple of unsafe business in this module, and some public
functions are unfortunately unsafe. Please refer to the inline comments
why some of them can be safe and some can't.

Thanks to Georges Racinet for the basic ideas and experiments,
          Raphaël Gomès for the core implementation,
          Mark Thomas for the generation-counter idea,
          and Mercurial developers for reviewing the original patches.