py_class: Add "@shared data" macro syntax
This mostly encapsulates PySharedRef business. The main reason of introducing
a dedicated macro syntax is that we need a safe way to bind the correct
PyObject reference to PySharedRefCell data. This is handled as follows:

    $crate::PySharedRef::new(py, &self._unsafe_inner, data)

Here `data` is a PySharedRefCell instance, and `self._unsafe_inner` is the
owner PyObject.

It's still weird that create_instance() takes PySharedRefCell<T> instead of
the inner data type T, which will be fixed by the next patch.